refactor: ♻️ move the ledger column list into its own module
The ledger presenter had grown past the size limit, so the column list
it shares with the table, the column picker and the exports now lives in
its own file and is re-exported from the presenter — no call site
changes. Ledger rows also gain the balance field the next commit fills.

Refs #2359
