+++ app/src/utils/accounting/ledgerPresenter.ts
@@ -17,6 +17,10 @@ import type { TokenId } from '@/constant'
 // Currency derivation / filtering lives in its own module, re-exported here.
 export { entryCurrency, ledgerCurrencies, filterLedgerByCurrency } from './ledgerCurrency'
 
+// So does the column list the table, the selector and the exporters share.
+export { LEDGER_COLUMNS, resolveLedgerColumns, ledgerTotalRow } from './ledgerColumns'
+export type { LedgerColumn, LedgerColumnKey } from './ledgerColumns'
+
 /** The empty activity carried by a posting's continuation (credit) and total rows. */
 const NO_ACTIVITY: ActivityCell = { kind: 'plain', text: '' }
 
@@ -69,71 +73,6 @@ export const ledgerCategories: Array<LedgerCategory | 'All' | typeof FEE_FILTER>
   FEE_FILTER
 ]
 
-/** The toggleable ledger table columns (keys match the table's cell slots). */
-export type LedgerColumnKey =
-  | 'date'
-  | 'action'
-  | 'transaction'
-  | 'activity'
-  | 'account'
-  | 'dr'
-  | 'cr'
-  | 'currency'
-  | 'quantity'
-  | 'rate'
-
-/** A ledger column as rendered in the selector and the exports. */
-export type LedgerColumn = { value: LedgerColumnKey; label: string }
-
-/**
- * Ledger columns as `{ value, label }`, for the show/hide-columns selector.
- * Devise / Quantité / Taux (spec §2) lead the USD debit/credit so each posting
- * reads "native currency · quantity · rate of record · $ moved".
- */
-export const LEDGER_COLUMNS: ReadonlyArray<LedgerColumn> = [
-  { value: 'date', label: 'Date' },
-  { value: 'action', label: 'Action' },
-  { value: 'transaction', label: 'Transaction' },
-  { value: 'activity', label: 'Activity' },
-  { value: 'account', label: 'Account' },
-  { value: 'currency', label: 'Currency' },
-  { value: 'quantity', label: 'Quantity' },
-  { value: 'rate', label: 'Rate' },
-  { value: 'dr', label: 'Debit' },
-  { value: 'cr', label: 'Credit' }
-]
-
-/**
- * The visible ledger columns for an export, in canonical order — an empty or
- * absent selection means "all columns". Shared by the PDF and Excel exporters so
- * both honour the same order regardless of the order columns were toggled.
- */
-export function resolveLedgerColumns(columns?: readonly LedgerColumnKey[]): LedgerColumn[] {
-  const visible = columns && columns.length ? columns : LEDGER_COLUMNS.map((c) => c.value)
-  return LEDGER_COLUMNS.filter((c) => visible.includes(c.value))
-}
-
-/**
- * The trailing "Total movements" row for an exported ledger, mirroring the
- * on-screen footer: the grand total in the Debit and Credit columns — as the
- * caller's already-rendered `amount` (a `$`-string for the PDF, a number for
- * Excel) — with the label in the Transaction column, or the first non-amount
- * column when Transaction is hidden.
- */
-export function ledgerTotalRow(
-  cols: readonly LedgerColumn[],
-  amount: string | number
-): (string | number)[] {
-  const labelKey = cols.some((c) => c.value === 'transaction')
-    ? 'transaction'
-    : cols.find((c) => c.value !== 'dr' && c.value !== 'cr')?.value
-  return cols.map((c) => {
-    if (c.value === 'dr' || c.value === 'cr') return amount
-    if (c.value === labelKey) return 'Total movements'
-    return ''
-  })
-}
-
 export interface LedgerRow {
   isFirst: boolean
   date: string
@@ -156,6 +95,9 @@ export interface LedgerRow {
   rate: string
   /** True on a `Transaction Fee Expense` leg — drives the "Fee" badge and filter. */
   isFee?: boolean
+  /** Running balance of the drilled account after this posting (see
+   *  {@link ./accountLedger.withRunningBalance}); absent outside a drill-down. */
+  balance?: string
 }
 
 export interface LedgerView {
